pyproject.toml
[tool.poetry]
name = "audio"
version = "0.1.0"
description = ""
authors = ["HeorhiiPotapov <bayjel13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^6.0.0"
ipython = "^8.8.0"
black = "^22.12.0"
isort = "^5.11.4"
scipy = {version = "^1.10.0", python = "3.10.9"}
python-dotenv = "^0.21.0"
matplotlib = "^3.6.2"
pandas = "^1.5.2"
tk = "^0.1.0"
llvmlite = {version = "^0.39.1", python = "3.10.9"}
numba = {version = "^0.56.4", python = "3.10.9"}
librosa = {version = "^0.9.2", python = "3.10.9"}
pyqt5 = "^5.15.7"
jupyter = "^1.0.0"
pygame = "^2.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.20.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
